requirements.txt
absl-py==0.15.0
antlr4-python3-runtime==4.8
anyio==3.3.1
appdirs==1.4.4
argcomplete==1.12.3
argon2-cffi==21.1.0
astor==0.8.1
attrs==21.2.0
Babel==2.9.1
backcall==0.2.0
beautifulsoup4==4.11.1
black==21.4b2
bleach==4.1.0
cached-property==1.5.2
cachetools==4.2.4
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
click==8.0.3
cloudpickle==2.0.0
cycler==0.10.0
Cython==0.29.24
debugpy==1.4.3
decorator==5.0.9
defusedxml==0.7.1
easydict==1.9
entrypoints==0.3
filelock==3.7.1
future==0.18.2
fvcore==0.1.5.post20211019
gast==0.2.2
gdown==4.5.1
google-auth==2.3.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.41.0
h5py==3.6.0
hydra-core==1.1.1
idna==3.2
importlib-metadata==4.8.1
importlib-resources==5.3.0
iopath==0.1.9
ipykernel==6.4.0
ipython==7.27.0
ipython-genutils==0.2.0
jedi==0.18.0
Jinja2==3.0.1
joblib==1.1.0
json5==0.9.6
jsonschema==3.2.0
jupyter-client==7.0.2
jupyter-core==4.7.1
jupyter-server==1.11.0
jupyterlab==3.1.11
jupyterlab-pygments==0.1.2
jupyterlab-server==2.8.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.4.3
matplotlib-inline==0.1.3
mistune==0.8.4
mypy-extensions==0.4.3
nbclassic==0.3.1
nbclient==0.5.4
nbconvert==6.1.0
nbformat==5.1.3
nest-asyncio==1.5.1
nltk==3.7
notebook==6.4.3
numpy==1.21.2
oauthlib==3.1.1
omegaconf==2.1.1
opencv-python==4.5.3.56
opt-einsum==3.3.0
packaging==21.0
pandas==1.3.2
pandocfilters==1.4.3
parso==0.8.2
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.3.2
portalocker==2.3.2
prometheus-client==0.11.0
prompt-toolkit==3.0.20
protobuf==3.18.1
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.2
pycparser==2.20
pydot==1.4.2
Pygments==2.10.0
pyparsing==2.4.7
pyrsistent==0.18.0
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2021.1
PyYAML==6.0
pyzmq==22.2.1
regex==2021.10.8
requests==2.26.0
requests-oauthlib==1.3.0
requests-unixsocket==0.2.0
rsa==4.7.2
scipy==1.1.0
seaborn==0.11.2
Send2Trash==1.8.0
six==1.16.0
sniffio==1.2.0
soupsieve==2.3.2.post1
tabulate==0.8.9
tensorboard==1.15.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow-estimator==1.15.1
tensorflow-gpu==1.15.2
termcolor==1.1.0
terminado==0.12.1
testpath==0.5.0
toml==0.10.2
torch==1.8.1
torchvision==0.9.1
tornado==6.1
tqdm==4.62.3
traitlets==5.1.0
typed-ast==1.4.3
typing-extensions==3.10.0.2
urllib3==1.26.6
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.2.1
Werkzeug==2.0.2
wrapt==1.13.3
yacs==0.1.8
zipp==3.5.0
